requirements.txt
numpy
pandas
streamlit
scikit-learn
python-docx
PyPDF2
pickle